feat: Daemon modularization + capability-gated command routing + CI/test pipeline updates
