+++ .github/actions/build-docs/action.yml
@@ -9,21 +9,9 @@ inputs:
 runs:
   using: 'composite'
   steps:
-    - name: Setup pnpm
-      uses: pnpm/action-setup@41ff72655975bd51cab0327fa583b6e92b6d3061 # v4.2.0
-      with:
-        version: 9.12.2
-
-    - name: Setup Node.js
-      uses: actions/setup-node@6044e13b5dc448c55e2357c09f80417699197238 # v6.2.0
-      with:
-        node-version: 22
-        cache: 'pnpm'
-
-    - name: Install dependencies (root)
+    - name: Setup toolchain
       if: ${{ github.event.action != 'closed' }}
-      run: pnpm install --frozen-lockfile
-      shell: bash
+      uses: ./.github/actions/setup-node-pnpm
 
     - name: Build docs
       if: ${{ github.event.action != 'closed' }}
