+++ .github/actions/setup-node-pnpm/action.yml
@@ -1,0 +1,35 @@
+name: 'Setup Node and pnpm'
+description: 'Setup pnpm, setup Node.js, and optionally install dependencies'
+
+inputs:
+  node-version:
+    description: 'Node.js version'
+    required: false
+    default: '22'
+  pnpm-version:
+    description: 'pnpm version'
+    required: false
+    default: '9.12.2'
+  install-deps:
+    description: 'Whether to install dependencies with pnpm install --frozen-lockfile'
+    required: false
+    default: 'true'
+
+runs:
+  using: 'composite'
+  steps:
+    - name: Setup pnpm
+      uses: pnpm/action-setup@41ff72655975bd51cab0327fa583b6e92b6d3061 # v4.2.0
+      with:
+        version: ${{ inputs.pnpm-version }}
+
+    - name: Setup Node.js
+      uses: actions/setup-node@6044e13b5dc448c55e2357c09f80417699197238 # v6.2.0
+      with:
+        node-version: ${{ inputs.node-version }}
+        cache: 'pnpm'
+
+    - name: Install dependencies
+      if: ${{ inputs.install-deps == 'true' }}
+      run: pnpm install --frozen-lockfile
+      shell: bash
