+++ skills/agent-device/SKILL.md
@@ -102,8 +102,8 @@ agent-device type "text"               # Type into focused field without clearin
 agent-device press 300 500             # Tap by coordinates
 agent-device long-press 300 500 800    # Long press (where supported)
 agent-device scroll down 0.5
-agent-device pinch 2.0              # Zoom in 2x (iOS simulator)
-agent-device pinch 0.5 200 400     # Zoom out at coordinates (iOS simulator)
+agent-device pinch 2.0              # Zoom in 2x (iOS simulator + Android)
+agent-device pinch 0.5 200 400     # Zoom out at coordinates
 agent-device back
 agent-device home
 agent-device app-switcher
@@ -141,7 +141,7 @@ agent-device apps --platform android --user-installed
 
 ## Best practices
 
-- Pinch (`pinch <scale> [x y]`) is supported on iOS simulators only; scale > 1 zooms in, < 1 zooms out.
+- Pinch (`pinch <scale> [x y]`) is supported on iOS simulators and Android; scale > 1 zooms in, < 1 zooms out. On Android, pinch uses multi-touch `sendevent` injection.
 - Always snapshot right before interactions; refs invalidate on UI changes.
 - Prefer `snapshot -i` to reduce output size.
 - On iOS, `xctest` is the default and does not require Accessibility permission.
