+++ src/core/dispatch.ts
@@ -17,6 +17,7 @@ import { setIosSetting } from '../platforms/ios/index.ts';
 import type { RawSnapshotNode } from '../utils/snapshot.ts';
 
 export type CommandFlags = {
+  session?: string;
   platform?: 'ios' | 'android';
   device?: string;
   udid?: string;
@@ -232,15 +233,11 @@ export async function dispatchCommand(
       if (Number.isNaN(scale) || scale <= 0) {
         throw new AppError('INVALID_ARGS', 'pinch requires scale > 0');
       }
-      if (device.platform === 'ios' && device.kind === 'simulator') {
-        await runIosRunnerCommand(
-          device,
-          { command: 'pinch', scale, x, y, appBundleId: context?.appBundleId },
-          { verbose: context?.verbose, logPath: context?.logPath, traceLogPath: context?.traceLogPath },
-        );
-      } else {
-        throw new AppError('UNSUPPORTED_OPERATION', 'pinch is only supported on iOS simulators');
-      }
+      await runIosRunnerCommand(
+        device,
+        { command: 'pinch', scale, x, y, appBundleId: context?.appBundleId },
+        { verbose: context?.verbose, logPath: context?.logPath, traceLogPath: context?.traceLogPath },
+      );
       return { scale, x, y };
     }
     case 'screenshot': {
@@ -251,9 +248,6 @@ export async function dispatchCommand(
     }
     case 'back': {
       if (device.platform === 'ios') {
-        if (device.kind !== 'simulator') {
-          throw new AppError('UNSUPPORTED_OPERATION', 'back is only supported on iOS simulators in v1');
-        }
         await runIosRunnerCommand(
           device,
           { command: 'back', appBundleId: context?.appBundleId },
@@ -266,9 +260,6 @@ export async function dispatchCommand(
     }
     case 'home': {
       if (device.platform === 'ios') {
-        if (device.kind !== 'simulator') {
-          throw new AppError('UNSUPPORTED_OPERATION', 'home is only supported on iOS simulators in v1');
-        }
         await runIosRunnerCommand(
           device,
           { command: 'home', appBundleId: context?.appBundleId },
@@ -281,9 +272,6 @@ export async function dispatchCommand(
     }
     case 'app-switcher': {
       if (device.platform === 'ios') {
-        if (device.kind !== 'simulator') {
-          throw new AppError('UNSUPPORTED_OPERATION', 'app-switcher is only supported on iOS simulators in v1');
-        }
         await runIosRunnerCommand(
           device,
           { command: 'appSwitcher', appBundleId: context?.appBundleId },
@@ -306,12 +294,6 @@ export async function dispatchCommand(
     case 'snapshot': {
       const backend = context?.snapshotBackend ?? 'xctest';
       if (device.platform === 'ios') {
-        if (device.kind !== 'simulator') {
-          throw new AppError(
-            'UNSUPPORTED_OPERATION',
-            'snapshot is only supported on iOS simulators in v1',
-          );
-        }
         if (backend === 'ax') {
           const ax = await snapshotAx(device, { traceLogPath: context?.traceLogPath });
           return { nodes: ax.nodes ?? [], truncated: false, backend: 'ax' };
