+++ src/daemon/context.ts
@@ -1,0 +1,34 @@
+import type { CommandFlags } from '../core/dispatch.ts';
+
+export function contextFromFlags(
+  logPath: string,
+  flags: CommandFlags | undefined,
+  appBundleId?: string,
+  traceLogPath?: string,
+): {
+  appBundleId?: string;
+  activity?: string;
+  verbose?: boolean;
+  logPath?: string;
+  traceLogPath?: string;
+  snapshotInteractiveOnly?: boolean;
+  snapshotCompact?: boolean;
+  snapshotDepth?: number;
+  snapshotScope?: string;
+  snapshotBackend?: 'ax' | 'xctest';
+  snapshotRaw?: boolean;
+} {
+  return {
+    appBundleId,
+    activity: flags?.activity,
+    verbose: flags?.verbose,
+    logPath,
+    traceLogPath,
+    snapshotInteractiveOnly: flags?.snapshotInteractiveOnly,
+    snapshotCompact: flags?.snapshotCompact,
+    snapshotDepth: flags?.snapshotDepth,
+    snapshotScope: flags?.snapshotScope,
+    snapshotRaw: flags?.snapshotRaw,
+    snapshotBackend: flags?.snapshotBackend,
+  };
+}
