+++ src/platforms/ios/index.ts
@@ -83,7 +83,7 @@ export async function closeIosApp(device: DeviceInfo, app: string): Promise<void
   ]);
 }
 
-export async function pressIos(device: DeviceInfo, x: number, y: number): Promise<void> {
+export async function pressIos(device: DeviceInfo, _x: number, _y: number): Promise<void> {
   ensureSimulator(device, 'press');
   throw new AppError(
     'UNSUPPORTED_OPERATION',
@@ -93,9 +93,9 @@ export async function pressIos(device: DeviceInfo, x: number, y: number): Promis
 
 export async function longPressIos(
   device: DeviceInfo,
-  x: number,
-  y: number,
-  durationMs = 800,
+  _x: number,
+  _y: number,
+  _durationMs = 800,
 ): Promise<void> {
   ensureSimulator(device, 'long-press');
   throw new AppError(
@@ -108,7 +108,7 @@ export async function focusIos(device: DeviceInfo, x: number, y: number): Promis
   await pressIos(device, x, y);
 }
 
-export async function typeIos(device: DeviceInfo, text: string): Promise<void> {
+export async function typeIos(device: DeviceInfo, _text: string): Promise<void> {
   ensureSimulator(device, 'type');
   throw new AppError(
     'UNSUPPORTED_OPERATION',
@@ -128,8 +128,8 @@ export async function fillIos(
 
 export async function scrollIos(
   device: DeviceInfo,
-  direction: string,
-  amount = 0.6,
+  _direction: string,
+  _amount = 0.6,
 ): Promise<void> {
   ensureSimulator(device, 'scroll');
   throw new AppError(
