+++ test/integration/smoke-cli.test.ts
